requirements.txt
PyYAML
js2py
curl_cffi